MINIFICPP-1177 Add new matchesFullInput() function
MINIFICPP-1177 Add file_size and computeChecksum

MINIFICPP-1177 Fix log rotation handling in TailFile

Instead of the dodgy logic of doing one file at a time in onTrigger
and trying to keep track using the persisted state, we follow the same
logic as NiFi now: only look at rolled files if the input file got
truncated, only look at new rotated files matching the pattern, and
stream all new content in one onTrigger call.

Also fixed some bugs in Multiple file mode and Yield detection logic,
as well as corrected some unit tests (when we checked the log output,
we previously included output from earlier stages of the test).

Support some previously unsupported NiFi properties:
- Recursive lookup
- Lookup frequency
- Rolling Filename Pattern

Change the default delimiter to \n (the previous default was to always
read to the current end of the file, even if it is in the middle of a
line).  Also include the delimiter in the flow file.  (As NiFi does.)

MINIFICPP-1177 Make TailFile unit tests more robust

These used to fail in a small percentage of cases due to timing issues.
I have reordered some operations and added some sleeps, and the failures
seem to be gone now.

MINIFICPP-1177 Use ProcessSession::write() instead of import()

MINIFICPP-1177 Make ProvenanceTestHelper thread-safe

Also fix variable naming, remove C-style casts, change virtual to override,
add missing overrides, and fix a bug in TestRepository::DeSerialize.

MINIFICPP-1177 Log the state at startup

Also remove some unused arguments and clean up some comments in the header.

MINIFICPP-1177 Save last_read_time to the state manager, too


This closes #791

+++ libminifi/src/core/ProcessSession.cpp
@@ -217,7 +217,9 @@ void ProcessSession::remove(const std::shared_ptr<core::FlowFile> &flow) {
   } else {
     logger_->log_debug("Flow does not contain content. no resource claim to decrement.");
   }
-  process_context_->getFlowFileRepository()->Delete(flow->getUUIDStr());
+  if (_addedFlowFiles.find(flow->getUUIDStr()) == _addedFlowFiles.end()) {
+    process_context_->getFlowFileRepository()->Delete(flow->getUUIDStr());
+  }
   _deletedFlowFiles[flow->getUUIDStr()] = flow;
   std::string reason = process_context_->getProcessorNode()->getName() + " drop flow record " + flow->getUUIDStr();
   provenance_report_->drop(flow, reason);
@@ -648,7 +650,20 @@ void ProcessSession::import(const std::string& source, std::vector<std::shared_p
 }
 
 void ProcessSession::import(std::string source, std::vector<std::shared_ptr<FlowFileRecord>> &flows, bool keepSource, uint64_t offset, char inputDelimiter) {
+// this function calls a deprecated function, but it is itself deprecated, so suppress warnings
+#if defined(__clang__)
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
+#elif defined(__GNUC__)
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wdeprecated-declarations"
+#endif
   import(source, flows, offset, inputDelimiter);
+#if defined(__clang__)
+#pragma clang diagnostic pop
+#elif defined(__GNUC__)
+#pragma GCC diagnostic pop
+#endif
   logger_->log_trace("Closed input %s, keeping source ? %i", source, keepSource);
   if (!keepSource) {
     std::remove(source.c_str());
@@ -970,6 +985,13 @@ bool ProcessSession::outgoingConnectionsFull(const std::string& relationship) {
     }
   }
   return false;
+}
+
+bool ProcessSession::existsFlowFileInRelationship(const Relationship &relationship) {
+  return std::any_of(_transferRelationship.begin(), _transferRelationship.end(),
+      [&relationship](const std::map<std::string, Relationship>::value_type &key_value_pair) {
+        return relationship == key_value_pair.second;
+  });
 }
 
 }  // namespace core
